refactor: improve the decoder mod
