+++ src/decoder/mod.rs
@@ -1,1 +1,3 @@
-pub mod image;
+mod image;
+
+pub use crate::decoder::image::Image;
